Add Saltbox to site configuration and reference it in navigation.
The MKDocs configuration was modified to include a new static site generator, Saltbox. This change was accomplished by modifying the 'extra' section of the configuration file. In addition, the navigation bar of the site was updated to include an item that references a newly added page that provides information about Saltbox.