Work tracking (#10)
* May 3, 2023, 9:40 AM

* May 3, 2023, 9:41 AM

* May 3, 2023, 9:43 AM

* May 3, 2023, 9:47 AM

* May 3, 2023, 9:52 AM

* May 3, 2023, 9:53 AM

* May 3, 2023, 1:58 PM

* May 3, 2023, 2:17 PM

* May 3, 2023, 2:18 PM

* May 3, 2023, 2:19 PM

* May 3, 2023, 2:19 PM

* yes

* May 3, 2023, 3:40 PM

* yessirski

* Update navigation bar to remove Docker documentation

Removes the Docker documentation from the navigation bar in testing section of the website. The Docker documentation can still be accessed through direct links.

* docs(code): Update Docker installation instructions

Updated the instructions for installing `Docker` to ensure the latest available version from the official `Docker` repository is used. The new installation method includes adding a package source, GPG key verification, and the addition of the `docker-ce` package. It also includes instructions for working with `Docker` images, such as checking if `Docker` is working correctly, searching for images available on Docker Hub, downloading an image to your computer, and listing all the downloaded images.

* Improve Docker tutorial content.

Added a link to Docker Hub and made minor text changes to our tutorial content for a better learning experience.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>